backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
cryptography>=42.0,<46.0
inflection>=0.5,<1.0
uritemplate>=4.1,<5.0
